bookmytea_auth/requirements.txt
bcrypt
pony
python-dotenv
pydantic
uvicorn
redis
pytest
requests
jsonrpcclient
flask
wtforms
pytelegrambotapi
fastapi-jsonrpc
psycopg2
pyjwt[crypto]